Spotify API call als frontend code
Ik heb de API call omgezet naar frontend code die nu in de browser draait. Het access token moet wel met client secret veilig worden opgehaald, daarom staat dat stukje code nog steeds in the backend, en wordt door middel van cors van de backend opgehaald
