+++ package.json
@@ -19,6 +19,7 @@
   "homepage": "https://github.com/sabrinaz35/team6#readme",
   "dependencies": {
     "bcryptjs": "^3.0.2",
+    "cors": "^2.8.5",
     "dotenv": "^16.4.7",
     "ejs": "^3.1.10",
     "express": "^4.21.2",
