+++ static/javascript/frontend.js
@@ -1,0 +1,57 @@
+console.log("Hallo wereld")
+
+//de access spotify access token van de backend ophalen
+async function getAccessToken(){
+    try{
+        const response = await fetch('http://localhost:4000/token');
+        const data = await response.json();
+        return data.access_token;
+    } catch(error){
+        console.error("Token not fetched", error)
+    }
+}
+
+//zoeken randomizen
+function getRandomSearch() {
+  // lijst van allemaal lowercase letters
+  const characters = 'abcdefghijklmnopqrstuvwxyz';
+  // haal een random letter uit deze lijst
+  const randomCharacter = characters.charAt(Math.floor(Math.random() * characters.length));
+  // deze letter returnen als zoek term
+  return randomCharacter;
+}
+
+
+// artist data van spotify opvragen
+async function fetchData() {
+  try {
+    // Access token opvragen voordat de data opgevraagd wordt
+    const accessToken = await getAccessToken(); 
+    //krijg een random array aan artiesten met een random begin letter
+    const response = await fetch(`https://api.spotify.com/v1/search?q=${getRandomSearch()}&type=artist&limit=50`, {
+      headers: {
+        Authorization: 'Bearer ' + accessToken
+      }
+    });
+
+    //alle artiesten data loggen
+    const data = await response.json();
+    console.log("Full artist data:", data.artists.items);
+
+    
+
+    //sorteer de array en check naar artiesten die populariteit =< 10 hebben
+    let kleineArtiesten = data.artists.items.filter(artist => artist.popularity <= 30);
+    console.log("kleine Artiesten zijn: " + JSON.stringify(kleineArtiesten, null, 2))
+
+  } catch (error) {
+    console.error('Error fetching data:', error);
+  }
+ 
+}
+
+// functie aanroepen
+fetchData();
+
+
+
