+++ views/partials/footer.ejs
@@ -1,0 +1,1 @@
+<script src="/static/javascript/frontend.js"></script>
